Add `python_source` config option
The `python_source` setting is really only relevant if the package in
question does not have a "standard" Python package layout, in which case
it may be difficult to infer/locate the source directory in which to
invoke `sphinx-apidoc`. This is particularly challenging for `mixed`
build types, so if it is specified by the user, this option can be used
as a fallback while locating the python source.

Signed-off-by: Abrar Rahman Protyasha <abrar@openrobotics.org>