Let `doxygen_builder` run if `run_doxygen` == True
Previously, `doxygen_builder` would only run if the build type was not
`ament_python`. This commit is part of a series of commits to get rid of
the `mixed` build type setting.

Instead, now, `doxygen_builder` is invoked either if the package build
type is `ament_cmake/cmake` or if `run_doxygen` is set to True.

Signed-off-by: Abrar Rahman Protyasha <abrar@openrobotics.org>