Unconditionally check `exhale`/`breathe` settings
In c6a6595, the `enable_exhale` and `enable_breathe` options were
evaluated only if the build type was `ament_cmake` or `cmake`.

Following the suggestion in
https://github.com/ros-infrastructure/rosdoc2/pull/28#discussion_r687043875,
this commit reverts that change, and modifies the behavior such that the
values of `enable_exhale` and `enable_breathe` are always checked. The
only difference now is that previously, the default value for these
settings was `True`. Now, the default is determined by what the build
type is and what the `always_run_doxygen` setting has been set to.

Essentially, this commit allows users to explicitly enable
`enable_exhale` or `enable_breathe` and have those extensions invoked
regardless of the build type.

Signed-off-by: Abrar Rahman Protyasha <abrar@openrobotics.org>